feat: change directionally controller into resolver (#278)
