+++ libs/template/application/root-app/src/root-app.component.ts
@@ -1,15 +1,11 @@
 import { LitRouter } from '@spryker-oryx/router/lit';
-import { hydratable, subscribe } from '@spryker-oryx/utilities';
+import { hydratable } from '@spryker-oryx/utilities';
 import { html, LitElement, TemplateResult } from 'lit';
-import { DirectionalityController } from './directionality.controller';
 import { styles } from './root-app.styles';
 
 @hydratable()
 export class RootAppComponent extends LitElement {
   static styles = styles;
-
-  @subscribe()
-  protected dirSetup = new DirectionalityController(this).install();
 
   protected router = new LitRouter(this, []);
