+++ libs/template/application/src/feature.ts
@@ -2,6 +2,7 @@ import { AppFeature } from '@spryker-oryx/core';
 import { isServer } from 'lit';
 import * as components from './components';
 import { RootPlugin } from './plugins';
+import { applicationProviders } from './services';
 export * from './components';
 
 export const applicationComponents = Object.values(components);
@@ -15,4 +16,5 @@ export const applicationFeature: AppFeature = {
       root: components.rootAppComponent,
     },
   },
+  providers: applicationProviders,
 };
