+++ libs/template/application/src/index.ts
@@ -2,3 +2,4 @@ export * from './auth';
 export * from './feature';
 export * from './plugins';
 export * from './ssr';
+export * from './services';
