+++ libs/template/application/src/services/resolvers/index.ts
@@ -1,0 +1,1 @@
+export * from './directionality-page-meta-resolver'
